Fix the zoom buttons in the pattern- and song editors to look the same, and fix implementation of zooming in the pattern editor.
git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@306 3b3fb362-3133-0410-aa15-cf69e0a59cb7

+++ gui/src/PatternEditor/DrumPatternEditor.h
@@ -67,7 +67,7 @@ class DrumPatternEditor : public QWidget, public EventListener, public Object
 		void updateEditor();
 
 	private:
-		uint m_nGridWidth;
+		float m_nGridWidth;
 		uint m_nGridHeight;
 		int m_nEditorHeight;
 		uint m_nResolution;
